requirements.txt
numpy
requests
Flask
pandas
scikit-learn
joblib
Selenium-Screenshot